задача 8

+++ 006/Program.cs
@@ -3,7 +3,7 @@
 
 Console.Clear();
 Console.WriteLine("Введите число:");
-double x = Convert.ToInt32(Console.ReadLine());
+int x = Convert.ToInt32(Console.ReadLine());
 if (x % 2 == 0)
     Console.WriteLine("Число является ЧЕТНЫМ");
 else
